pyproject.toml
[tool.poetry]
name = "orkg-simcomp-api"
version = "0.1.0"
description = "REST API to ORKG-SimComp"
authors = ["Yaser Jaradeh <jaradeh@l3s.de>", "Omar Arab Oghli <omar.araboghli@outlook.com>"]
maintainers = ["Yaser Jaradeh", "Omar Arab Oghli"]
readme = "README.md"
homepage = "http://orkg.org/about"
repository = "https://gitlab.com/TIBHannover/orkg/orkg-simcomp/orkg-simcomp-api/"
keywords = ["ORKG", "SimComp", "Scholarly Communication", "Digital Library"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8,<3.10"
fastapi = "~0.92.0"
gunicorn = "~20.1.0"
uvicorn = "~0.17.6"
python-dotenv = "0.20.0"
SQLAlchemy = "^1.4.39"
psycopg2-binary = "2.9.3"
requests = "2.28.1"
orkg = "0.15.0"
elasticsearch = "8.5.3"
BaseHash = "3.0.10"
pre-commit = "^3.1.0"
httpx = "^0.23.3"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"
parameterized = "0.8.1"

[tool.black]
line-length = 99

[tool.isort]
profile = "black"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
